requirements.txt
streamlit>=1.34.0
pandas>=2.2.0
plotly>=5.22.0
python-pptx>=0.6.23
openpyxl>=3.1.2
